LayoutTests/platform/glib/imported/w3c/web-platform-tests/notifications/permissions-non-secure-expected.txt

CONSOLE MESSAGE: The Notification permission may only be requested in a secure context.

PASS Notification.requestPermission must be called from a secure context
PASS Notification.permission must be called from a secure context
